move network commands to network namespace
