+++ lib/api/client-commands/network/captureNetworkRequests.js
@@ -1,5 +1,5 @@
-const ClientCommand = require('./_base-command.js');
-const {Logger} = require('../../utils');
+const ClientCommand = require('../_base-command.js');
+const {Logger} = require('../../../utils/index.js');
 
 /**
  * Capture outgoing network calls from the browser.
@@ -27,6 +27,10 @@ const {Logger} = require('../../utils');
  * @moreinfo nightwatchjs.org/guide/network-requests/capture-network-calls.html
  */
 class CaptureNetworkCalls extends ClientCommand {
+
+  static get namespacedAliases() {
+    return 'captureNetworkRequests';
+  }
 
   performAction(callback) {
