+++ lib/api/client-commands/network/mockNetworkResponse.js
@@ -1,5 +1,5 @@
-const ClientCommand = require('./_base-command.js');
-const {Logger} = require('../../utils');
+const ClientCommand = require('../_base-command.js');
+const {Logger} = require('../../../utils/index.js');
 
 /**
  * Intercept the request made on a particular URL and mock the response.
@@ -30,6 +30,10 @@ const {Logger} = require('../../utils');
  * @moreinfo nightwatchjs.org/guide/network-requests/mock-network-response.html
  */
 class MockNetworkResponse extends ClientCommand {
+
+  static get namespacedAliases() {
+    return 'mockNetworkResponse';
+  }
 
   performAction(callback) {
