+++ lib/api/client-commands/network/setNetworkConditions.js
@@ -1,5 +1,5 @@
-const ClientCommand = require('./_base-command.js');
-const {Logger} = require('../../utils');
+const ClientCommand = require('../_base-command.js');
+const {Logger} = require('../../../utils/index.js');
 
 /**
  * 
@@ -23,6 +23,11 @@ const {Logger} = require('../../utils');
  * @api protocol.sessions
  */
 class SetNetworkConditions extends ClientCommand {
+
+  static get namespacedAliases() {
+    return 'setNetworkConditions';
+  }
+
   performAction(callback) {
     if (!this.api.isChrome() && !this.api.isEdge()) {
       const error = new Error('SetNetworkConditions is not supported while using this driver');
